backend/fashion-search/requirements.txt
boto3>=1.28.0
psycopg2-binary>=2.9.0
requests>=2.31.0